Correct length in host IP and port extraction
The latest build for The QT6WIP didnt connect to a socketcand server, from the console output i noticed that the port it tried to connect was :0 instead of the actual port. I compared the differences between the main and qt6 branches and noticed this difference in the ip and port extraction part, changing it and building for a couple hours fixed the issue. I also build it myself with the original value and that did not work, im confident this is the proper fix for this issue.